safec filename changes + pulledpork3.conf
Had to change how the safec library is downloaded and handled by tar. jq parses out all of the browser download urls for the latest release, and looks for the line that ends with ".tar.bz2" in order to download the correct package now. We also are using tar -xjvf with the -C and --strip-components=1 option to make a directory name that is at least somewhat consistent.

I also added a VERY stripped down pulledpork3.conf file to repo. Its own my TODO list to switch to pulledpork3.